io: move public types out of io::driver.
The `Interest` and `Ready` types are part of the public API and should
not be defined in the driver module, which is intended to be internal.

This is in preparation to move the I/O driver into the runtime module.
