+++ tokio/src/signal/unix/driver.rs
@@ -2,8 +2,8 @@
 
 //! Signal driver
 
-use crate::io::driver::{Driver as IoDriver, Interest};
-use crate::io::PollEvented;
+use crate::io::driver::Driver as IoDriver;
+use crate::io::{Interest, PollEvented};
 use crate::park::Park;
 use crate::signal::registry::globals;
